added IdentityUser + database. Will need configuration
Currently using decimal for numbers. Needs mapping for database because of this warning: "No store type was specified for the decimal property 'Amount' on entity type 'Expense'. This will cause values to be silently truncated if they do not fit in the default precision and scale". Counts for all decimal variables types in model