TryGetUserLookupIdWithValidation
Validate's the user & gets the correct user id. A request has a optional userId. This is used when a user wants to edit a different users models. Passes the user that the person wants to edit from and this method checks if they are both valid and if the current user is authorized to edit